Update patch set 1
Patch Set 1: Code-Review+1

The ifdefs and flags work fine with compiling tinycompress, but the errors now occur when we build the audio HAL. We either need to redeclare the flag there, or just add it to build/core/qcom_target.mk and have it declared for everything. Save time and just go with the latter method. See http://review.cyanogenmod.org/#/c/96771/

Patch-set: 1
Label: Code-Review=+1